Aqua refactor: Move all enums under `enums.py` for centralized management (#848)
