+++ ads/aqua/common/enums.py
@@ -11,8 +11,14 @@ This module contains the set of enums used in AQUA.
 from ads.common.extended_enum import ExtendedEnumMeta
 
 
+class DataScienceResource(str, metaclass=ExtendedEnumMeta):
+    MODEL_DEPLOYMENT = "datasciencemodeldeployment"
+    MODEL = "datasciencemodel"
+
+
 class Resource(str, metaclass=ExtendedEnumMeta):
     JOB = "jobs"
+    JOBRUN = "jobruns"
     MODEL = "models"
     MODEL_DEPLOYMENT = "modeldeployments"
     MODEL_VERSION_SET = "model-version-sets"
@@ -21,3 +27,28 @@ class Resource(str, metaclass=ExtendedEnumMeta):
 class DataScienceResource(str, metaclass=ExtendedEnumMeta):
     MODEL_DEPLOYMENT = "datasciencemodeldeployment"
     MODEL = "datasciencemodel"
+
+
+class Tags(str, metaclass=ExtendedEnumMeta):
+    TASK = "task"
+    LICENSE = "license"
+    ORGANIZATION = "organization"
+    AQUA_TAG = "OCI_AQUA"
+    AQUA_SERVICE_MODEL_TAG = "aqua_service_model"
+    AQUA_FINE_TUNED_MODEL_TAG = "aqua_fine_tuned_model"
+    AQUA_MODEL_NAME_TAG = "aqua_model_name"
+    AQUA_EVALUATION = "aqua_evaluation"
+    AQUA_FINE_TUNING = "aqua_finetuning"
+    READY_TO_FINE_TUNE = "ready_to_fine_tune"
+    READY_TO_IMPORT = "ready_to_import"
+    BASE_MODEL_CUSTOM = "aqua_custom_base_model"
+
+
+class RqsAdditionalDetails(str, metaclass=ExtendedEnumMeta):
+    METADATA = "metadata"
+    CREATED_BY = "createdBy"
+    DESCRIPTION = "description"
+    MODEL_VERSION_SET_ID = "modelVersionSetId"
+    MODEL_VERSION_SET_NAME = "modelVersionSetName"
+    PROJECT_ID = "projectId"
+    VERSION_LABEL = "versionLabel"
