+++ ads/aqua/evaluation/evaluation.py
@@ -26,7 +26,12 @@ from oci.data_science.models import (
 from ads.aqua import logger
 from ads.aqua.app import AquaApp
 from ads.aqua.common import utils
-from ads.aqua.common.enums import *
+from ads.aqua.common.enums import (
+    DataScienceResource,
+    Resource,
+    RqsAdditionalDetails,
+    Tags,
+)
 from ads.aqua.common.errors import (
     AquaFileExistsError,
     AquaFileNotFoundError,
@@ -44,7 +49,6 @@ from ads.aqua.common.utils import (
     is_valid_ocid,
     upload_local_to_os,
 )
-from ads.aqua.data import Tags
 from ads.aqua.evaluation.constants import *
 from ads.aqua.evaluation.entities import *
 from ads.aqua.evaluation.errors import *
